Add multiple morph support and demonstrate possible bug of missing CONTACT_END messages
This enhances the contacts example to handle multiple Morphs, and
demonstrates a possible bug that causes CONTACT_END messages to be
omitted or lost.  The LEDs are turned on/off when CONTACT_START and
CONTACT_END messages are received.  It also changes it so that it runs
forever, until you hit controlC.  If you run this with 4 Morphs, and tap
one of the Morphs with 3 fingers, quickly, it will often leave one or
more of the LEDS lit.  Looking at the messages that are printed, you can
see that one or more of the contacts doesn't send a CONTACT_END message.
It seems to happen even with 1 Morph connected, so I don't think this
bug is multi-Morph-related, but enhancing this example to handle
multiple Morphs (or creating a new example to handle multiple Morphs)
would be nice to include in the examples, independent of how the bug (if
it is a bug) is fixed.